move worldsync messages to folder
